fix: 修改人均次数

+++ co-driver-statistics/co-driver-statistics-services/co-driver-statistics-service-ads/src/main/java/com/leo/ad/codriver/ads/entity/AdsDailyOetaBaseReport.java
@@ -45,9 +45,9 @@ public class AdsDailyOetaBaseReport {
     private Long adDirectSoldClickNum;
     private Long adDirectSoldClickShowRate;
     private BigDecimal adDirectSoldRate;
-    private Long avgUserAdShowNum;
-    private Long avgUserAdInterShowNum;
-    private Long avgUserAdMrecShowNum;
+    private BigDecimal avgUserAdShowNum;
+    private BigDecimal avgUserAdInterShowNum;
+    private BigDecimal avgUserAdMrecShowNum;
     private BigDecimal withdrawAmount;
     private BigDecimal withdrawFee;
     private BigDecimal withdrawCost;
